restructured application layer
