+++ src/core/application/mail_service.ts
@@ -24,11 +24,3 @@ export interface IMailService {
      */
     startEmailSequence(sequenceType: MailJobType, data: any): void;
 }
-
-export interface ILoggerService {
-    silly(message: string, meta?: any): void;
-    error(message: string, meta?: any): void;
-    info(message: string, meta?: any): void;
-    debug(message: string, meta?: any): void;
-    warn(message: string, meta?: any): void;
-}
