+++ src/infrastructure/services/mail_service.ts
@@ -1,7 +1,7 @@
 import Agenda from "agenda";
 import Mailgun from "mailgun-js";
 import { TYPES } from "../../core/domain/constants/types";
-import { IMailService } from "../../core/application/services";
+import { IMailService } from "../../core/application/mail_service";
 import { config } from "../config";
 import { iocContainer, provideSingleton } from "../config/ioc";
 import { MailJobType } from "../jobs/mail_job";
