--- src/ui/service/user_service.ts
@@ -1,19 +1,0 @@
-import { UserSignUpInput } from "../models/user_dto";
-import { User } from "../../core/domain/models/user";
-
-export interface IUserService {
-    create(user: UserSignUpInput): Promise<User>;
-    get(query: { id?: string; emailOrUsername?: string }): Promise<User>;
-    getAll(): Promise<User[]>;
-    pagedGetAll({
-        searchStr,
-        skip,
-        limit
-    }: {
-        searchStr?: string;
-        skip?: number;
-        limit?: number;
-    }): Promise<{ totalCount: number; items: User[] }>;
-    update(user: Partial<User>): Promise<void>;
-    delete(id: string): Promise<boolean>;
-}
