+++ src/ui/subscribers/user.ts
@@ -1,6 +1,7 @@
 import { EventSubscriber, On } from "event-dispatch";
 import mongoose from "mongoose";
-import { ILoggerService, IMailService } from "../../core/application/services";
+import { ILoggerService } from "../../core/application/logger_service";
+import { IMailService } from "../../core/application/mail_service";
 import { User } from "../../core/domain/models/user";
 import { iocContainer } from "../../infrastructure/config/ioc";
 import { MailJobType } from "../../infrastructure/jobs/mail_job";
